requirements.txt
rdkit
matplotlib
pytorch
pandas
numpy
pytorch-lightning
joblib
tqdm
PyTDC
chemprop
tensorboard
